move import image command  to PromptMine Connector and implemented the command
